# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateWorkflowDefinitionRequest:
    properties:
      category:
        description: 流程分类
        example: dev
        type: string
      definitionKey:
        description: 流程标识，系统内唯一
        example: story_approval
        type: string
      definitionName:
        description: 流程名称
        example: 需求审批流程
        type: string
      flowData:
        description: LogicFlow画布JSON
        example: '{"nodes":[],"edges":[]}'
        type: string
      remark:
        description: 备注
        example: 流程说明
        type: string
    required:
    - definitionKey
    - definitionName
    type: object
...
  models.Response:
    properties:
      code:
        description: 状态码 0=成功 -1=失败
        example: 0
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
      message:
        description: 提示信息
        example: ok
        type: string
    type: object
...
  models.UpdateWorkflowCanvasRequest:
    properties:
      flowData:
        description: LogicFlow画布JSON
        example: '{"nodes":[],"edges":[]}'
        type: string
    required:
    - flowData
    type: object
  models.UpdateWorkflowDefinitionRequest:
    properties:
      category:
        description: 流程分类
        example: dev
        type: string
      definitionKey:
        description: 流程标识，系统内唯一
        example: story_approval
        type: string
      definitionName:
        description: 流程名称
        example: 需求审批流程
        type: string
      flowData:
        description: LogicFlow画布JSON
        example: '{"nodes":[],"edges":[]}'
        type: string
      remark:
        description: 备注
        example: 流程说明
        type: string
    required:
    - definitionKey
    - definitionName
    type: object
  models.UpdateWorkflowFormSchemaRequest:
    properties:
      formSchemaId:
        example: UUID
        type: string
    required:
    - formSchemaId
    type: object
  models.UpdateWorkflowStatusRequest:
    properties:
      status:
        description: 流程状态：0草稿 1已发布 2已停用
        example: "2"
        type: string
    required:
    - status
    type: object
...
  models.WorkflowDefinitionResponse:
    properties:
      category:
        description: 流程分类
        example: dev
        type: string
      createDate:
        description: 创建时间
        example: "2026-05-18 15:30:26"
        type: string
      creatorId:
        description: 创建人ID
        example: UUID
        type: string
      creatorName:
        description: 创建人姓名
        example: 管理员
        type: string
      definitionId:
        description: 流程定义ID
        example: UUID
        type: string
      definitionKey:
        description: 流程标识
        example: story_approval
        type: string
      definitionName:
        description: 流程名称
        example: 需求审批流程
        type: string
      flowData:
        description: LogicFlow画布JSON
        example: '{"nodes":[],"edges":[]}'
        type: string
      formSchemaId:
        description: 关联表单Schema ID
        example: UUID
        type: string
      remark:
        description: 备注
        example: 流程说明
        type: string
      status:
        description: 流程状态：0草稿 1已发布 2已停用
        example: "0"
        type: string
      updateDate:
        description: 更新时间
        example: "2026-05-18 15:30:26"
        type: string
      version:
        description: 发布版本号
        example: 1
        type: integer
    type: object
...
  utils.PaginationResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
host: localhost:9191
info:
  contact: {}
  description: Hive Admin Go 后端 API 接口文档
  title: Hive Admin API
  version: "1.0"
paths:
...
  /workflow/definitions:
    delete:
      consumes:
      - application/json
      description: 批量删除流程定义
      parameters:
      - description: 流程定义ID数组
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除流程定义
      tags:
      - 流程管理-流程定义
    get:
      consumes:
      - application/json
      description: 分页获取流程定义列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 流程标识
        in: query
        name: definitionKey
        type: string
      - description: 流程名称
        in: query
        name: definitionName
        type: string
      - description: 流程分类
        in: query
        name: category
        type: string
      - description: 流程状态，支持多选：0,1,2
        in: query
        name: status
        type: string
      - description: 排序参数，支持：definitionKey、definitionName、category、status、version、createDate、updateDate
        in: query
        name: sorts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/utils.PaginationResponse'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/models.WorkflowDefinitionResponse'
                        type: array
                    type: object
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取流程定义列表
      tags:
      - 流程管理-流程定义
    post:
      consumes:
      - application/json
      description: 创建新的流程定义
      parameters:
      - description: 流程定义信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateWorkflowDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建流程定义
      tags:
      - 流程管理-流程定义
  /workflow/definitions/{definitionId}:
    get:
      consumes:
      - application/json
      description: 根据流程定义ID获取流程定义详情
      parameters:
      - description: 流程定义ID
        in: path
        name: definitionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.WorkflowDefinitionResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取流程定义详情
      tags:
      - 流程管理-流程定义
    put:
      consumes:
      - application/json
      description: 更新流程定义基础信息和画布数据
      parameters:
      - description: 流程定义ID
        in: path
        name: definitionId
        required: true
        type: string
      - description: 流程定义信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateWorkflowDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新流程定义
      tags:
      - 流程管理-流程定义
  /workflow/definitions/{definitionId}/canvas:
    put:
      consumes:
      - application/json
      description: 保存 LogicFlow 画布JSON数据
      parameters:
      - description: 流程定义ID
        in: path
        name: definitionId
        required: true
        type: string
      - description: 流程画布信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateWorkflowCanvasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 保存流程画布
      tags:
      - 流程管理-流程定义
  /workflow/definitions/{definitionId}/formSchema:
    put:
      consumes:
      - application/json
      description: 保存流程定义关联的表单 Schema ID，并将流程恢复为草稿状态
      parameters:
      - description: 流程定义ID
        in: path
        name: definitionId
        required: true
        type: string
      - description: 表单 Schema ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateWorkflowFormSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数或表单配置错误
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 绑定流程表单 Schema
      tags:
      - 流程管理-流程定义
  /workflow/definitions/{definitionId}/publish:
    put:
      consumes:
      - application/json
      description: 发布流程定义并递增版本号
      parameters:
      - description: 流程定义ID
        in: path
        name: definitionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 发布成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 发布流程定义
      tags:
      - 流程管理-流程定义
  /workflow/definitions/{definitionId}/status:
    put:
      consumes:
      - application/json
      description: 更新流程定义状态：0草稿 1已发布 2已停用
      parameters:
      - description: 流程定义ID
        in: path
        name: definitionId
        required: true
        type: string
      - description: 状态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateWorkflowStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无接口访问权限
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新流程定义状态
      tags:
      - 流程管理-流程定义
  /workflow/definitions/all:
    get:
      consumes:
      - application/json
      description: 获取所有流程定义（不分页）
      parameters:
      - description: 流程名称
        in: query
        name: definitionName
        type: string
      - description: 流程分类
        in: query
        name: category
        type: string
      - description: 流程状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.WorkflowDefinitionResponse'
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有流程定义
      tags:
      - 流程管理-流程定义
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"